% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcbn.R
\name{ct_cbn}
\alias{ct_cbn}
\title{Calls ct-cbn}
\usage{
ct_cbn(filestem, N = 0, e = 0, B = NULL, M = 0, R = 1, S = 1)
}
\description{
`ct_cbn` performs the operations on the files.
}
\details{
This is text that will be replaced, or forgotten
}
