% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initializes package}
\usage{
init()
}
\description{
`init` initalizes all of the executables needed for the package to run.
}
\details{
This function should be called only once, right after the package is installed.
Requires gsl and gcc to be installed. Both can be installed on Mac using brew install and on alternate operating systems through corresponding functions.
}
